LayoutTests/platform/wpe/http/tests/xmlhttprequest/onloadend-event-after-error-expected.txt

CONSOLE MESSAGE: Load cannot follow more than 20 redirections
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8000/xmlhttprequest/resources/infinite-loop.php due to access control checks.
Test case for bug 40952: Onloadend event is not supported in XMLHttpRequest

Verify that a loadend ProgressEvent is dispatched after the error ProgressEvent when an async request fails.

Verify also that a progress ProgressEvent is sent between readystatechange and error ProgressEvent (bug 126575).

PASS should appear below:

PASS
